pkgAcquire::Worker::CurrentSize was changed to unsigned long long
[mspang: Added cast]

+++ contrib/acqprogress.cc
@@ -217,7 +217,7 @@ bool AcqTextStatus::Pulse(pkgAcquire *Owner)
             
       // Add the current progress
       if (Mode == Long)
-	 snprintf(S,End-S," %lu",I->CurrentSize);
+	 snprintf(S,End-S," %llu",(unsigned long long)I->CurrentSize);
       else
       {
 	 if (Mode == Medium || I->TotalSize == 0)
